Move utils to AzTest
