+++ Code/Framework/AzTest/AzTest/Platform/Android/platform_android_files.cmake
@@ -8,6 +8,7 @@
 
 set(FILES
     ../Common/UnixLike/AzTest/ColorizedOutput_UnixLike.cpp
+    ../Common/Unimplemented/AzTest/Utils_Unimplemented.cpp
     ScopedAutoTempDirectory_Android.cpp
     Platform_Android.cpp
     AzTest_Traits_Platform.h
