+++ Code/Framework/AzTest/AzTest/Platform/Mac/platform_mac_files.cmake
@@ -9,6 +9,7 @@
 set(FILES
     ../Common/UnixLike/AzTest/ColorizedOutput_UnixLike.cpp
     ../Common/UnixLike/AzTest/ScopedAutoTempDirectory_UnixLike.cpp
+    ../Common/UnixLike/AzTest/Utils_UnixLike.cpp
     Platform_Mac.cpp
     AzTest_Traits_Platform.h
     AzTest_Traits_Mac.h
