+++ Code/Framework/AzTest/AzTest/Platform/iOS/platform_ios_files.cmake
@@ -9,6 +9,7 @@
 set(FILES
     ../Common/UnixLike/AzTest/ColorizedOutput_UnixLike.cpp
     ../Common/Unimplemented/AzTest/ScopedAutoTempDirectory_Unimplemented.cpp
+    ../Common/Unimplemented/AzTest/Utils_Unimplemented.cpp
     Platform_iOS.cpp
     AzTest_Traits_Platform.h
     AzTest_Traits_iOS.h
